+++ DotML.Cli/src/Embeddings/BaseEmbedding.cs
@@ -8,9 +8,9 @@ namespace DotML.Cli.Embeddings;
 /// An embedding that must be provided by a file and not accessible via standard input
 /// </summary>
 public abstract class FileOnlyEmbedding : IEmbedder {
-    public abstract BatchedFeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files);
+    public abstract BatchedFeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files);
 
-    public BatchedFeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, string raw) {
+    public BatchedFeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, string raw) {
         throw new NotSupportedException($"{GetType()} embedding doesn't support data from stdin.");
     }
 }
@@ -20,12 +20,12 @@ public abstract class FileOnlyEmbedding : IEmbedder {
 /// </summary>
 public abstract class ImageEmbedding : FileOnlyEmbedding {
 
-    public override BatchedFeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files) {
+    public override BatchedFeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, IEnumerable<FileInfo> files) {
         var batches = files.Select(file => CreateEmbedding(@for, file)).ToArray();
-        return new BatchedFeatureSet<double>(batches);
+        return new BatchedFeatureSet<float>(batches);
     }
 
-    public FeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, FileInfo file) {
+    public FeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, FileInfo file) {
         using var original = SKBitmap.FromImage(SKImage.FromEncodedData(file.FullName));
         using var processed = PreprocessImage(@for, original);
         var features = CreateEmbedding(@for, processed);
@@ -70,9 +70,9 @@ public abstract class ImageEmbedding : FileOnlyEmbedding {
         return resized;
     }
 
-    public abstract FeatureSet<double> CreateEmbedding(FeedforwardNetwork @for, SKBitmap bitmap);
+    public abstract FeatureSet<float> CreateEmbedding(FeedforwardNetwork @for, SKBitmap bitmap);
 
-    public virtual FeatureSet<double> Postprocess(FeatureSet<double> features) {
+    public virtual FeatureSet<float> Postprocess(FeatureSet<float> features) {
         return features;
     }
