+++ DotML.Cli/src/Logging/IOutputLogger.cs
@@ -3,4 +3,4 @@ using DotML.Network.Training;
 
 namespace DotML.Cli.Logging;
 
-public interface IOutputLogger : ILayerInputVisitor<(int LayerIndex, BatchedFeatureSet<double> Output)> { }
+public interface IOutputLogger : ILayerInputVisitor<(int LayerIndex, BatchedFeatureSet<float> Output)> { }
