+++ DotML/src/NeuralNetwork/INeuron.cs
@@ -8,12 +8,12 @@ public interface INeuron {
     /// Neuron bias value
     /// </summary>
     /// <value>bias</value>
-    public double Bias {get; set;}
+    public float Bias {get; set;}
     /// <summary>
     /// Neuron input weights
     /// </summary>
     /// <value>vector of weights, one per input</value>
-    public Span<double> Weights {get;}
+    public Span<float> Weights {get;}
     /// <summary>
     /// Activation function for outputs
     /// </summary>
